Password length control has been turned into a function to make it more controllable.
After the password was created, the user was asked whether he wanted to create a password again. It has been designed in such a way that the program will run again if desired, if not, the program will terminate, and if an incorrect entry is made, the same question will be asked to the user again.